New feature : + Changed the BaseApiController interface to IHttpActionResult
Bug fixed :
- Fixed the Domain.Customer attribute MembershipTypeId type from 'byte' to 'long'

+++ Vidly.Data/Domain/Customer.cs
@@ -16,7 +16,7 @@ namespace Vidly.Domain
         public bool IsSubscribedToNewsletter { get; set; }
 
         [Column("MEMBERSHIPTYPEID")]
-        public byte MembershipTypeId { get; set; }
+        public long MembershipTypeId { get; set; }
 
         [Column("BIRTHDATE")]
         public Nullable<DateTime> Birthdate { get; set; }
